requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.13.0
async-timeout==5.0.1
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
distro==1.9.0
exceptiongroup==1.3.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.3
google-auth==2.49.2
google-cloud-core==2.5.1
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-genai==1.73.1
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langextract==1.2.1
MarkupSafe==3.0.3
ml_collections==1.1.0
more-itertools==11.0.2
multidict==6.7.1
narwhals==2.19.0
numpy==2.4.4
packaging==26.1
pandas==3.0.2
pillow==12.2.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==7.34.1
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.2
pydantic_core==2.46.2
pydeck==0.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
streamlit==1.56.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
watchdog==6.0.0
websockets==16.0
yarl==1.23.0
